+++ plend/src/main/java/three/people/service/UserService.java
@@ -10,5 +10,6 @@ public interface UserService {
 	int idCheck(String id);
 	int nickNameCheck(String nickName);
 	UserVO selectID(UserVO vo);
+	public int tempPwd(UserVO vo);
 	public int selectPwd(UserVO vo);
 }
