+++ plend/src/main/webapp/WEB-INF/views/common/searchId.jsp
@@ -40,10 +40,10 @@
 			data : sendData,
 			dataType : "text",
 			success : function(text) {
-				if (text != null) {
+				if (text != "") {
 					$("#searchedId").html("<h2>등록된 아이디는  " + text +" 입니다.</h2>");
 					console.log(text);
-				} else{
+				} else if (text == ""){
 					$("#searchedId").html("등록된 정보가 없습니다.");
 				}
 			},
